+++ Examples/Resources/IntuneMobileAppsBundleMacOS/3-Remove.ps1
@@ -27,6 +27,7 @@ Configuration Example
         IntuneMobileAppsBundleMacOS "IntuneMobileAppsBundleMacOS-Example"
         {
             DisplayName           = "macOS Pkg App";
+            PackageFileType       = "Pkg";
             Ensure                = "Absent";
             ApplicationId         = $ApplicationId;
             TenantId              = $TenantId;